self-guided-labs/industrial-use-cases/orchestrate/travel/Voice_Agent/agent_session/requirements.txt
# FastAPI and Server
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart
pydantic>=2.0.0
pydantic-settings>=2.0.0

# IBM Watson & WatsonX
pyarrow>=14.0.1
ibm-watson>=6.1.0
ibm-cloud-sdk-core>=3.16.0
ibm-watsonx-ai>=1.0.0

# Machine Learning
torch>=2.0.0
transformers>=4.30.0

# Utils
python-dotenv>=0.19.0
pyyaml>=6.0
dateparser>=1.0.0
websockets>=10.0
requests>=2.28.0

# Audio Processing
soundfile>=0.12.0
librosa>=0.10.0
accelerate>=0.26.0